requirements.txt
flask==2.3.3
flask-cors==4.0.0
firecrawl-py==2.16.5
requests==2.31.0
python-dotenv==1.0.0 